server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
aiohttp==3.9.5
httpx==0.27.0
apscheduler==3.10.4
ultralytics==8.2.0
Pillow==10.3.0
python-dotenv==1.0.1
firebase-admin==6.5.0
python-multipart==0.0.9
